+++ core/src/stored/bextract.cc
@@ -96,10 +96,7 @@ static uint64_t fileAddr = 0; /* file write address */
 
 int main(int argc, char* argv[])
 {
-  setlocale(LC_ALL, "");
-  tzset();
-  bindtextdomain("bareos", LOCALEDIR);
-  textdomain("bareos");
+  INIT_LANGUAGES("");
   InitStackDump();
 
   working_directory = "/tmp";