+++ CHANGELOG.md
@@ -74,6 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - dird: remove unused write verify list [PR #2720]
 - core: add machine readable version to client hello; deprecate TlsAuthenticate [PR #2752]
 - core: move complete authentication into the library; remove support for old consoles (<18.2) [PR #2765]
+- core: remove setlocale initialization [PR #2819]
 
 ### Fixed
 - VMware Plugin: Fix NVRAM backup when datacenter is not in root folder [PR #2461]
@@ -2402,6 +2403,7 @@ If you want to migrate from your manually configured disk autochanger to simply 
 [PR #2810]: https://github.com/bareos/bareos/pull/2810
 [PR #2813]: https://github.com/bareos/bareos/pull/2813
 [PR #2814]: https://github.com/bareos/bareos/pull/2814
+[PR #2819]: https://github.com/bareos/bareos/pull/2819
 [PR #2827]: https://github.com/bareos/bareos/pull/2827
 [PR #2831]: https://github.com/bareos/bareos/pull/2831
 [PR #2833]: https://github.com/bareos/bareos/pull/2833